+++ src/internet_identity/src/activity_stats/activity_counter.rs
